requirements.txt
streamlit==1.32.0
python-docx==1.1.0
PyPDF2==3.0.1
langchain==0.1.12
chromadb==0.4.24
openai==1.12.0
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
tiktoken==0.6.0 